fix(html): make radio group the base primitive
