+++ packages/html/src/ui/menu/context.ts
@@ -13,11 +13,6 @@ export interface MenuContextValue {
   parentMenu: MenuApi | null;
 }
 
-export interface MenuRadioGroupContextValue {
-  value: string;
-  onValueChange: (value: string) => void;
-}
-
 export interface MenuGroupContextValue {
   registerLabel: (id: string) => () => void;
 }
@@ -29,13 +24,16 @@ export interface MenuItemSettingContextValue {
 }
 
 const MENU_CONTEXT_KEY = Symbol('@videojs/menu');
-const MENU_RADIO_GROUP_CONTEXT_KEY = Symbol('@videojs/menu-radio-group');
 const MENU_GROUP_CONTEXT_KEY = Symbol('@videojs/menu-group');
 const MENU_ITEM_SETTING_CONTEXT_KEY = Symbol('@videojs/menu-item-setting');
 
 export const menuContext = createContext<MenuContextValue>(MENU_CONTEXT_KEY);
-export const menuRadioGroupContext = createContext<MenuRadioGroupContextValue>(MENU_RADIO_GROUP_CONTEXT_KEY);
 export const menuGroupContext = createContext<MenuGroupContextValue>(MENU_GROUP_CONTEXT_KEY);
 export const menuItemSettingContext = createContext<MenuItemSettingContextValue | undefined>(
   MENU_ITEM_SETTING_CONTEXT_KEY
 );
+
+export {
+  radioGroupContext as menuRadioGroupContext,
+  type RadioGroupContextValue as MenuRadioGroupContextValue,
+} from '../radio-group/context';
