+++ packages/html/src/ui/menu/menu-radio-group-element.ts
@@ -1,33 +1,64 @@
-import type { PropertyDeclarationMap, PropertyValues } from '@videojs/element';
-import { ContextProvider } from '@videojs/element/context';
+import { resolveTranslation, type Translator } from '@videojs/core/i18n';
+import type { PropertyValues } from '@videojs/element';
 
-import { MediaElement } from '../media-element';
-import { menuRadioGroupContext } from './context';
+import { RadioGroupElement } from '../radio-group/radio-group-element';
 import { MenuGroupController } from './menu-group-controller';
+import { MenuRadioItemElement } from './menu-radio-item-element';
 
-export class MenuRadioGroupElement extends MediaElement {
+export class MenuRadioGroupElement extends RadioGroupElement {
   static readonly tagName: string = 'media-menu-radio-group';
 
-  static override properties = {
-    value: { type: String },
-  } satisfies PropertyDeclarationMap<'value'>;
-
-  value = '';
-
-  readonly #provider = new ContextProvider(this, { context: menuRadioGroupContext });
   readonly #group = new MenuGroupController(this);
+  #ariaLabel: string | null = null;
 
-  protected override update(_changed: PropertyValues): void {
-    super.update(_changed);
+  protected override update(changed: PropertyValues): void {
+    super.update(changed);
 
     this.#group.applyProps();
+  }
 
-    this.#provider.setValue({
-      value: this.value,
-      onValueChange: (next: string) => {
-        this.value = next;
-        this.dispatchEvent(new CustomEvent('value-change', { detail: { value: next }, bubbles: true }));
-      },
-    });
+  protected getTemplate(): HTMLTemplateElement | null {
+    for (const child of this.children) {
+      if (child instanceof HTMLTemplateElement) return child;
+    }
+
+    return null;
+  }
+
+  protected createRadioItem(template: HTMLTemplateElement | null): MenuRadioItemElement {
+    if (!template) return document.createElement(MenuRadioItemElement.tagName) as MenuRadioItemElement;
+
+    const fragment = template.content.cloneNode(true) as DocumentFragment;
+    const root = fragment.firstElementChild;
+
+    if (!root || root.localName !== MenuRadioItemElement.tagName || root.nextElementSibling) {
+      return document.createElement(MenuRadioItemElement.tagName) as MenuRadioItemElement;
+    }
+
+    return root as MenuRadioItemElement;
+  }
+
+  protected setItemLabel(item: MenuRadioItemElement, label: string): void {
+    const labelPart = item.querySelector<HTMLElement>('[data-part~="label"]');
+
+    if (labelPart) {
+      labelPart.textContent = label;
+    } else {
+      item.textContent = label;
+    }
+  }
+
+  protected applyAriaLabel(
+    translator: Translator,
+    label: string,
+    params?: Record<string, string | number>
+  ): void {
+    if (this.hasAttribute('aria-labelledby')) return;
+
+    const current = this.getAttribute('aria-label');
+    if (current !== null && current !== this.#ariaLabel) return;
+
+    this.#ariaLabel = resolveTranslation(translator, label, params);
+    this.setAttribute('aria-label', this.#ariaLabel);
   }
 }
