+++ packages/html/src/ui/quality-radio-group/quality-radio-group-element.ts
@@ -8,8 +8,8 @@ import { I18nController } from '../../i18n/controller';
 import { playerContext } from '../../player/context';
 import { PlayerController } from '../../player/player-controller';
 import { MenuItemIndicatorElement } from '../menu/menu-item-indicator-element';
+import { MenuRadioGroupElement } from '../menu/menu-radio-group-element';
 import { MenuRadioItemElement } from '../menu/menu-radio-item-element';
-import { RadioGroupElement } from '../radio-group/radio-group-element';
 
 function resolveAutoLabel(translator: Translator, label: string): string {
   const match = /^Auto \((.+)\)$/.exec(label);
@@ -19,11 +19,11 @@ function resolveAutoLabel(translator: Translator, label: string): string {
   return resolveTranslation(translator, label);
 }
 
-export class QualityRadioGroupElement extends RadioGroupElement {
+export class QualityRadioGroupElement extends MenuRadioGroupElement {
   static override readonly tagName = 'media-quality-radio-group';
 
   static override properties = {
-    ...RadioGroupElement.properties,
+    ...MenuRadioGroupElement.properties,
     disabled: { type: Boolean },
     label: { type: String },
   } satisfies PropertyDeclarationMap<'value' | 'label' | 'disabled'>;
